pyproject.toml
[tool.poetry]
name = "coin-users"
version = "0.1.0"
description = ""
authors = ["hwangtate <ghkdxodud10253504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
black = "^24.8.0"
isort = "^5.13.2"
coverage = "^7.6.1"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
python-dotenv = "^1.0.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
